platforms: move platforms to applications
The platform directory is a bit messy.
A platform is the integration of some base component like motion control,
avoidance, planning, ...
However in the actual state:
* platforms tends to be application code
* real platform code is under common directory
To correct that:
* platforms become applications
* common is removed and its content move into platforms
An application is now relying on a platform.
