+++ applications/cogip2019-cortex-simulation/main.c
