+++ applications/cogip2019-cortex/Makefile
