+++ applications/cogip2019-cortex/include/app.h
@@ -8,7 +8,7 @@
 
 /* Project includes */
 #include "board.h"
-#include "platform-common.h"
+#include "platform.h"
 #include "sd21.h"
 #include "vl53l0x.h"
