+++ applications/cogip2019-cortex/main.c
@@ -1,7 +1,6 @@
 #include <stdio.h>
 
 #include "platform.h"
-#include "platform-common.h"
 
 int main(void)
 {
