+++ platforms/cortex/include/platform.h
